requirements.txt
absl-py
langdetect
nltk
immutabledict
openai
transformers